backend/requirements.txt
google-generativeai
fastapi
uvicorn
crewai
duckduckgo-search
langchain_google_genai
setuptools
python-dotenv==1.0.1